Added a bunch of aliases and a function.
Added an alias for ls: it now correctly lists the contents of
symbolically linked directories and also has the appended characters
that classify files.

Added a bunch of aliases for make misspellings.

Added suffix aliases. Basically if you type a file name that ends in a
certain suffix, if you press enter you'll open the associated program.

Added comments.

Added keep function. Pipe and function output to keep, and then use the
variable $kept to access that command's output without having to re-type
it.